add two simple tests for state transitions, and more cleanup
